# backend/openapi/v1.yaml
openapi: 3.1.0
info:
  title: TDP API
  version: 1.0.0
  description: Unified API for TDP Lite (Go backend, single tenant)
servers:
  - url: https://api.example.com
security:
  - TDPHmac: []
components:
  securitySchemes:
    TDPHmac:
      type: apiKey
      in: header
      name: X-TDP-Signature
      description: |
        Required headers:
        - X-TDP-Key-Id
        - X-TDP-Timestamp
        - X-TDP-Nonce
        - X-TDP-Signature
  headers:
    Idempotency-Key:
      schema:
        type: string
        maxLength: 128
  schemas:
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        retryable: { type: boolean }
        requestId: { type: string }
    Locale:
      type: string
      enum: [en, zh]
    ContentStatus:
      type: string
      enum: [draft, published, archived]
    ContentKind:
      type: string
      enum: [post, moment, gallery]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
    AiProvider:
      type: string
      enum: [openai, anthropic, gemini]
    Post:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        locale: { $ref: '#/components/schemas/Locale' }
        title: { type: string }
        excerpt: { type: string, nullable: true }
        content: { type: string }
        coverUrl: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/ContentStatus' }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        revision: { type: integer }
    Moment:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        locale: { $ref: '#/components/schemas/Locale' }
        visibility:
          type: string
          enum: [public, private]
        media:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [image, video] }
              url: { type: string }
              width: { type: integer, nullable: true }
              height: { type: integer, nullable: true }
              thumbnailUrl: { type: string, nullable: true }
        location:
          type: object
          nullable: true
          properties:
            name: { type: string }
            lat: { type: number, nullable: true }
            lng: { type: number, nullable: true }
        status: { $ref: '#/components/schemas/ContentStatus' }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GalleryItem:
      type: object
      properties:
        id: { type: string }
        locale: { $ref: '#/components/schemas/Locale' }
        fileUrl: { type: string }
        thumbUrl: { type: string, nullable: true }
        title: { type: string, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        capturedAt: { type: string, format: date-time, nullable: true }
        camera: { type: string, nullable: true }
        lens: { type: string, nullable: true }
        focalLength: { type: string, nullable: true }
        aperture: { type: string, nullable: true }
        iso: { type: integer, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        isLivePhoto: { type: boolean }
        videoUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ContentStatus' }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AIJob:
      type: object
      properties:
        id: { type: string }
        kind: { $ref: '#/components/schemas/ContentKind' }
        contentId: { type: string }
        provider: { $ref: '#/components/schemas/AiProvider' }
        model: { type: string }
        prompt: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        result:
          type: object
          nullable: true
          additionalProperties: true
    Presence:
      type: object
      properties:
        online: { type: boolean }
        status: { type: string, enum: [online, offline, unknown] }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        country: { type: string, nullable: true }
        countryCode: { type: string, nullable: true }
        timezone: { type: string, nullable: true }
        source: { type: string, nullable: true }
        locationLabel: { type: string }
        lastHeartbeatAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        staleAfterSec: { type: integer, nullable: true }
    ProfileSnapshot:
      type: object
      properties:
        github:
          type: object
          nullable: true
          additionalProperties: true
        music:
          type: object
          nullable: true
          additionalProperties: true
        derived:
          type: object
          nullable: true
          additionalProperties: true
        sourceStatus:
          type: object
          nullable: true
          additionalProperties: true
        syncedAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
paths:
  /healthz:
    get:
      security: []
      responses:
        '200': { description: OK }
  /readyz:
    get:
      security: []
      responses:
        '200': { description: Ready }
        '503': { description: Not ready }
  /v1/public/feed:
    get:
      security: []
      parameters:
        - in: query
          name: locale
          schema: { $ref: '#/components/schemas/Locale' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Feed items }
  /v1/public/posts:
    get:
      security: []
      responses: { '200': { description: Post list } }
  /v1/public/posts/{slug}:
    get:
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses: { '200': { description: Post detail }, '404': { description: Not found } }
  /v1/public/moments:
    get:
      security: []
      responses: { '200': { description: Moment list } }
  /v1/public/moments/{id}:
    get:
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: Moment detail }, '404': { description: Not found } }
  /v1/public/gallery:
    get:
      security: []
      responses: { '200': { description: Gallery list } }
  /v1/public/gallery/{id}:
    get:
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: Gallery detail }, '404': { description: Not found } }
  /v1/public/presence:
    get:
      security: []
      responses:
        '200': { description: Presence status }
  /v1/public/profile-snapshot:
    get:
      security: []
      responses:
        '200': { description: Profile snapshot }
  /v1/public/search-snapshot:
    get:
      security: []
      parameters:
        - in: query
          name: locale
          schema: { $ref: '#/components/schemas/Locale' }
      responses:
        '200': { description: Search snapshot }
  /v1/public/search:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [section, query, locale]
              properties:
                section: { type: string, enum: [post, moment, gallery] }
                query: { type: string, minLength: 2, maxLength: 200 }
                locale: { $ref: '#/components/schemas/Locale' }
                cursor: { type: string }
                limit: { type: integer, minimum: 1, maximum: 30 }
                filters:
                  type: object
                  properties:
                    localeScope: { type: string, enum: [all, current] }
                    dateFrom: { type: string, pattern: '^\\d{4}-\\d{2}-\\d{2}$' }
                    dateTo: { type: string, pattern: '^\\d{4}-\\d{2}-\\d{2}$' }
                    tags:
                      type: array
                      items: { type: string }
                    location: { type: string }
                    camera: { type: string }
                    lens: { type: string }
                    focalLength: { type: string }
                    aperture: { type: string }
                    isoMin: { type: integer, minimum: 0 }
                    isoMax: { type: integer, minimum: 0 }
      responses:
        '200': { description: Section search results }
  /v1/internal/presence:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [city]
              properties:
                city: { type: string, minLength: 1, maxLength: 120 }
                region: { type: string, maxLength: 120 }
                country: { type: string, maxLength: 120 }
                countryCode: { type: string, maxLength: 8 }
                timezone: { type: string, maxLength: 80 }
                source: { type: string, maxLength: 80 }
      responses:
        '200': { description: Presence updated }
  /v1/internal/profile-snapshot:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                github:
                  type: object
                  additionalProperties: true
                music:
                  type: object
                  additionalProperties: true
                derived:
                  type: object
                  additionalProperties: true
                sourceStatus:
                  type: object
                  additionalProperties: true
                syncedAt:
                  type: string
                  format: date-time
      responses:
        '200': { description: Profile snapshot updated }
  /v1/internal/search-snapshot:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200': { description: Search snapshot updated }
  /v1/internal/search-snapshot/status:
    get:
      responses:
        '200': { description: Search snapshot refresh status }
  /v1/media/uploads:
    post:
      parameters:
        - $ref: '#/components/headers/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimeType, size]
              properties:
                filename: { type: string }
                mimeType: { type: string }
                size: { type: integer }
                sha256: { type: string }
      responses:
        '200': { description: Upload created }
  /v1/media/uploads/{uploadId}/complete:
    post:
      parameters:
        - in: path
          name: uploadId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Upload complete }
  /v1/previews/sessions:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId: { type: string }
                payload:
                  type: object
                  additionalProperties: true
                kind: { $ref: '#/components/schemas/ContentKind' }
                contentId: { type: string }
      responses:
        '200': { description: Preview session }
  /v1/previews/sessions/{id}/payload:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: exp
          required: true
          schema: { type: string }
        - in: query
          name: sig
          required: true
          schema: { type: string }
      responses:
        '200': { description: Preview payload }
  /v1/posts:
    post:
      parameters:
        - $ref: '#/components/headers/Idempotency-Key'
      responses: { '200': { description: Create post } }
  /v1/posts/{id}:
    patch:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: Update post } }
    delete:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: Delete post } }
  /v1/posts/{id}/publish:
    post:
      responses: { '200': { description: Publish post } }
  /v1/posts/{id}/unpublish:
    post:
      responses: { '200': { description: Unpublish post } }
  /v1/moments:
    post:
      responses: { '200': { description: Create moment } }
  /v1/moments/{id}:
    patch:
      responses: { '200': { description: Update moment } }
    delete:
      responses: { '200': { description: Delete moment } }
  /v1/moments/{id}/publish:
    post:
      responses: { '200': { description: Publish moment } }
  /v1/moments/{id}/unpublish:
    post:
      responses: { '200': { description: Unpublish moment } }
  /v1/gallery-items:
    post:
      responses: { '200': { description: Create gallery item } }
  /v1/gallery-items/{id}:
    patch:
      responses: { '200': { description: Update gallery item } }
    delete:
      responses: { '200': { description: Delete gallery item } }
  /v1/gallery-items/{id}/publish:
    post:
      responses: { '200': { description: Publish gallery item } }
  /v1/gallery-items/{id}/unpublish:
    post:
      responses: { '200': { description: Unpublish gallery item } }
  /v1/ai/jobs:
    post:
      responses: { '200': { description: Create AI job } }
  /v1/ai/jobs/{jobId}:
    get:
      responses: { '200': { description: AI job detail } }
  /v1/ai/jobs/{jobId}/apply:
    post:
      responses: { '200': { description: Apply AI suggestion } }
  /v1/ai/models:
    get:
      responses: { '200': { description: AI models } }
  /v1/keys:
    get:
      responses: { '200': { description: Key list } }
    post:
      responses: { '200': { description: Create key } }
  /v1/keys/{id}/rotate:
    post:
      responses: { '200': { description: Rotate key } }
  /v1/keys/{id}/revoke:
    post:
      responses: { '200': { description: Revoke key } }
  /v1/jobs/{id}:
    get:
      responses: { '200': { description: Generic job status } }